services/post-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
aio-pika>=9.4.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
aiokafka>=0.10.0
redis[hiredis]>=5.0.0
cassandra-driver>=3.29.0
